dynamic_memory_eval/requirements.txt
sentence-transformers
torch
PyYAML
requests
